allow start with an initial coeff_m_bit_count
This is useful when we know a correct parameters set, we can use it as
initial values and we can directly test (evaluate) with it. Testing will
give us noise growth. The remainding budget at the end of the evaluation
will allow us directly to infer minimal correct parameters
Allow specifying a min xdepth (if we are not interested in testing with
smaller xdepths)